+++ soh/src/overlays/actors/ovl_En_Ishi/z_en_ishi.c
@@ -316,7 +316,7 @@ void EnIshi_Init(Actor* thisx, PlayState* play) {
         this->actor.uncullZoneForward += 1000.0f;
     }
     if (this->actor.shape.rot.y == 0) {
-        this->actor.shape.rot.y = this->actor.world.rot.y = Rand_ZeroFloat(0x10000);
+        this->actor.shape.rot.y = this->actor.world.rot.y = (s16)(u16)(s32)Rand_ZeroFloat(0x10000);
     }
     Actor_SetScale(&this->actor, sRockScales[type]);
     EnIshi_InitCollider(&this->actor, play);
